Submitting the Coupon Form
So let's move to phase two: when we hit 'Add', this should submit to a new endpoint that validates the code in Stripe and attaches it to our user's cart.

To create the new endpoint, open OrderController. Near the bottom add a new public function addCouponAction() with @Route('/checkout/coupon'). Name it order_add_coupon. And to be extra-hipster, add @Method('POST') to guarantee that you can only POST to this.
Cool! Copy the route name, then find the coupon form in the checkout template. Update the form's action: add path() and then paste the route name.
Next, we'll read the submitted code and check with Stripe to make sure it's real, and not just someone trying to guess clever coupon codes. Come on, we've all tried it before.