Fetching the Coupon Information
Great! At this point, all we need to do is talk to Stripe and ask them:

Hey Stripe! Is this a valid coupon code in your system?. Oh, and if it is, ow much is it for?
Since Coupon is just an object in Stripe's API, we can fetch it like anything else. Booya!

As usual, add the API call in StripeClient. At the bottom, create a new public function called findCoupon() with a {dollar_sign}code argument. Then, return \Stripe\Coupon::retrieve() and pass it the  string, which is the Coupon's primary key in Stripe's API.