Back in OrderController, add {dollar_sign}stripeCoupon = {dollar_sign}this->get('stripe_client') and then call ->findCoupon({dollar_sign}code). If the code is invalid, Stripe will throw an exception. We'll handle that in a few minutes. But just for now, let's dump({dollar_sign}stripeCoupon) and die to see what it looks like.
Ok, refresh, hit 'I have a coupon code,' fill in our CHEAP_SHEEP code, and submit!

There it is! In the _values section where the data hides, the coupon has an id, it shows the amount_off in cents and has a few other things, like duration, in case you want to create coupons that are recurring and need to tell the user that this will be applied multiple times.