There are two ways to use a coupon on checkout: either attach it to the subscription to say 'This subscription should have this coupon code' - or - attach it to the customer. They're approximately the same, but we'll attach the coupon to the customer, in part, because the coupon should also work on individual products.
In OrderController, scroll down to the chargeCustomer() method.
We know this method: we get or create the Stripe Customer, create InvoiceItems for any products, create the Subscription, and then create an invoice, if needed.

Before adding the invoice items, let's add the coupon to the Customer. So, if {dollar_sign}cart->getCouponCodeValue(), then very simply, {dollar_sign}stripeCustomer->coupon = {dollar_sign}cart->getCouponCode(). Make it official with {dollar_sign}customer->save().

The important thing is that you don't need to change how much you're charging the user: attach the coupon, charge them for the full amount, and let Stripe figure it all out.

I think we should try this out! Use our favorite fake credit card, and Checkout! So far so good!

Find the Customer in Stripe. Yep! There's the order: 9. The invoice tells the whole story: with the sub-total, the discount and the total.

Very, very, nice.