Handling Invalid Coupons
And very easy! So easy, that we have time to add code to handle invalid coupons. Add another item to your cart. Now, try a FAKE coupon code.

Ah! 500 error is no fun. The exception is a \Stripe\Error\InvalidRequest because, basically, the API responds with a 404 status code.

This all falls apart in OrderController on line 95. Hunt that down!

Ah, findCoupon(): surround this beast with a try-catch block for \Stripe\Error\InvalidRequest
The easiest thing to do is add a flash error message: Invalid Coupon code. Then, redirect back to the checkout page.
Refresh that bad coupon! Ok! That's covered!