Go create a huge coupon - like for USD500. Call this one SUPER_SHEEP!
Ok, this coupon is awesome - so let's add it to an order.

Perfect! As you can see, the cart is already smart enough to return the total as USD0, instead of a negative number. Way to go cart!

But, uh, what's this checkout form still doing over here? Why should I need to enter my credit card info? This order is free!

Look, how you handle free orders is up to you. If you still want to require a credit card, you can do that. The customer won't be charged, but the card will be on file for renewals.

But that's not for me: I want to make it as easy as possible for this user to checkout free.

Hide the Checkout Form

The first step, is to hide this checkout form! In checkout.html.twig, find the _cardForm.html.twig include, and wrap it in if cart.totalWithDiscount > 0. Else, create a really simple form that submits right back to this URL. Give it method='POST' and a submit button that invites the user to 'Checkout for Free'

If you refresh now, boom! Checkout form gone.