Handling Free Checkout
But that's not quite everything we need to do. Thanks to the discount, Stripe will already know that it doesn't need to charge the user, and so, the customer doesn't need to have a card. But, the funny thing is, up until now, our OrderController logic is expecting a stripeToken to always be submitted:
169 lines  src/AppBundle/Controller/OrderController.php
... lines 1 - 11
class OrderController extends BaseController
{
... lines 14 - 48
  public function checkoutAction(Request {dollar_sign}request)
  {
... lines 51 - 53
    if ({dollar_sign}request->isMethod('POST')) {
      {dollar_sign}token = {dollar_sign}request->request->get('stripeToken');
... lines 56 - 68
    }
... lines 70 - 77
  }
... lines 79 - 166
}
... lines 168 - 169
Remember that's the token that we get back from Stripe, after submitting the credit card information to them. We then pass that to chargeCustomer() and attach it to the Customer:

 169 lines  src/AppBundle/Controller/OrderController.php
... lines 1 - 11
class OrderController extends BaseController
{
... lines 14 - 120
  private function chargeCustomer()
  {
... lines 123 - 125
  if (! {dollar_sign}user->getStripeCustomerId()) {
    {dollar_sign}stripeCustomer = {dollar_sign}stripeClient->createCustomer({dollar_sign}user, {dollar_sign}token);
  } else {
      {dollar_sign}stripeCustomer = {dollar_sign}stripeClient->updateCustomerCard({dollar_sign}user, {dollar_sign}token);
    }
... lines 131 - 165
  }
}

Optionally Apply the Stripe Token

But now, our code needs to be smart enough to not try to attach the token to the Customer for free orders. At the top, add a sanity check: if there is no token, and the shopping cart's total with discount is not free... well, we have a problem! Throw a clear exception: the order is non-free... but we're missing the payment token!