Ok, I'm feeling good! The last trouble spot is in updateCardDetails(). Open SubscriptionHelper. Oh yea - this method looks at the sources key on the Customer to get the card brand and last four digits. In our app, every Customer has exactly one card, so we use the 0 key. But guess what! Not anymore: a customer might have zero cards.
So we just need to code defensively: add an if statement: if ! {dollar_sign}stripeCustomer->sources->data, just return: there's no card details to update.

Ok, we're done! A free checkout and a normal checkout are almost the same. The only difference is that you don't have a Stripe token, so you can't attach that to your Customer.

Refresh the checkout page and, 'Checkout for Free'.

It works! In Stripe, find our Customer. There is no new payment, but there is an Invoice for USD0 and an active subscription. The Invoice shows off the discount.