Ha! We've made it! We've survived our subscription payment setup! So, umm, go celebrate: eat some cake! Sing a song! Or, like we'll do, do some accounting.
Because our last topic is about that: your users will need to see a receipt after purchase. And good news: we've setup our system so that every charge is done by creating an Invoice in Stripe's system. That'll make our life easy.

Store Invoices Locally?

In fact, all the information we need to render a receipt is already stored in Stripe. So, you could create a local invoices database table and store details there... or, you can take a shortcut and use Stripe's API to fetch invoice data whenever you actually need it. Let's do that.